requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
identify==2.6.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
openai==1.40.1
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240807
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
praw==7.7.1
prawcore==2.4.0
pre-commit==3.8.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
SciencePlots==2.1.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.5
transformers==4.44.0
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
